publisher-requirements.txt
boto3>=1.37,<2
google-api-python-client
google-auth
tenacity>=8.0